Added a singleton TagColorMap to be shared across PersonCard and PersonInfoPanel
