+++ src/main/java/seedu/address/ui/PersonCard.java
@@ -1,7 +1,5 @@
 package seedu.address.ui;
 
-import java.util.HashMap;
-
 import javafx.beans.binding.Bindings;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
@@ -17,12 +15,7 @@ import seedu.address.model.person.ReadOnlyPerson;
 public class PersonCard extends UiPart<Region> {
 
     private static final String FXML = "PersonListCard.fxml";
-    private static final String[] COLORS = {"Crimson", "orange", "DarkSalmon", "LightSeaGreen",
-        "RoyalBlue", "MediumPurple", "Teal", "Sienna", "HotPink", "MediumSeaGreen",
-        "DarkSlateBlue"};
-    private static final int NUM_COLORS = COLORS.length;
-    private static int colorIndex = 0;
-    private static HashMap<String, String> tagColors = new HashMap<String, String>();
+
     public final ReadOnlyPerson person;
 
 
@@ -57,30 +50,6 @@ public class PersonCard extends UiPart<Region> {
     }
 
     /**
-     * Gets a random unused color for the new tagName, or returns the corresponding color of the old tagName
-     * @param tagName
-     * @return the color of the tag
-     */
-    private String getTagColor(String tagName) {
-        if (!tagColors.containsKey(tagName)) {
-            tagColors.put(tagName, COLORS[colorIndex]);
-            updateColorIndex();
-        }
-        return tagColors.get(tagName);
-    }
-
-    /**
-     * Updates the color index to pick a new color for the new tag.
-     */
-    private static void updateColorIndex() {
-        if (colorIndex == NUM_COLORS - 1) {
-            colorIndex = 0;
-        } else {
-            colorIndex++;
-        }
-    }
-
-    /**
      * Binds the individual UI elements to observe their respective {@code Person} properties
      * so that they will be notified of any changes.
      */
@@ -101,7 +70,7 @@ public class PersonCard extends UiPart<Region> {
     private void initTags(ReadOnlyPerson person) {
         person.getTags().forEach(tag -> {
                 Label tagLabel = new Label(tag.tagName);
-                tagLabel.setStyle("-fx-background-color: " + getTagColor(tag.tagName));
+                tagLabel.setStyle("-fx-background-color: " + TagColorMap.getInstance().getTagColor(tag.tagName));
                 tags.getChildren().add(tagLabel);
             }
         );
