+++ src/main/java/seedu/address/ui/PersonInfoPanel.java
@@ -1,7 +1,6 @@
 package seedu.address.ui;
 
 import java.io.File;
-import java.util.HashMap;
 import java.util.logging.Logger;
 
 import com.google.common.eventbus.Subscribe;
@@ -31,16 +30,9 @@ public class PersonInfoPanel extends UiPart<Region> {
     private static final String FXML = "PersonInfoPanel.fxml";
     private static String DEFAULT_PHOTO_PATH = "/images/defaultPhoto.jpg";
 
-    private static final String[] COLORS = {"Crimson", "orange", "DarkSalmon", "LightSeaGreen",
-        "RoyalBlue", "MediumPurple", "Teal", "Sienna", "HotPink", "MediumSeaGreen",
-        "DarkSlateBlue"};
-    private static final int NUM_COLORS = COLORS.length;
-    private static int colorIndex = 0;
-
     private ReadOnlyPerson person;
 
     private final Logger logger = LogsCenter.getLogger(this.getClass());
-    private HashMap<String, String> tagColors = new HashMap<String, String>();
 
     @FXML
     private Circle photoCircle;
@@ -120,34 +112,10 @@ public class PersonInfoPanel extends UiPart<Region> {
     private void initTags(ReadOnlyPerson person) {
         person.getTags().forEach(tag -> {
                 Label tagLabel = new Label(tag.tagName);
-                tagLabel.setStyle("-fx-background-color: " + getTagColor(tag.tagName));
+                tagLabel.setStyle("-fx-background-color: " + TagColorMap.getInstance().getTagColor(tag.tagName));
                 tags.getChildren().add(tagLabel);
             }
         );
-    }
-
-    /**
-     * Gets a random unused color for the new tagName, or returns the corresponding color of the old tagName
-     * @param tagName
-     * @return the color of the tag
-     */
-    private String getTagColor(String tagName) {
-        if (!tagColors.containsKey(tagName)) {
-            tagColors.put(tagName, COLORS[colorIndex]);
-            updateColorIndex();
-        }
-        return tagColors.get(tagName);
-    }
-
-    /**
-     * Updates the color index to pick a new color for the new tag.
-     */
-    private static void updateColorIndex() {
-        if (colorIndex == NUM_COLORS - 1) {
-            colorIndex = 0;
-        } else {
-            colorIndex++;
-        }
     }
 
     //@@author April0616
